requirements.txt
uvicorn 
fastapi
pydantic
pandas 
plotly
matplotlib
shap
numpy
scipy
imblearn
lightgbm
scikit-optimize
sklearn
boruta
streamlit
request